Move the kwargs filter helper into models/callbacks
