+++ airflow-core/src/airflow/utils/helpers.py
@@ -18,7 +18,6 @@
 from __future__ import annotations
 
 import copy
-import inspect
 import itertools
 import re
 import signal
@@ -294,29 +293,6 @@ def prune_dict(val: Any, mode="strict"):
                 new_list.append(v)
         return new_list
     return val
-
-
-def filter_kwargs(callable_obj: Callable[..., Any], kwargs: dict) -> dict:
-    """
-    Filter kwargs to only include parameters the callable accepts.
-
-    If the callable accepts **kwargs (VAR_KEYWORD), all kwargs are passed through.
-    Otherwise, only kwargs matching named parameters are passed.  This is useful
-    when calling user-provided callables that may not accept all of the kwargs that
-    Airflow injects (e.g. context).
-
-    :param callable_obj: The callable to inspect
-    :param kwargs: The full set of kwargs to filter
-    """
-    try:
-        signature = inspect.signature(callable_obj)
-    except (ValueError, TypeError):
-        return kwargs
-
-    if any(param.kind == inspect.Parameter.VAR_KEYWORD for param in signature.parameters.values()):
-        return kwargs
-
-    return {k: v for k, v in kwargs.items() if k in set(signature.parameters.keys())}
 
 
 __deprecated_imports = {
